Allow appliction to be loaded with JTAG
And not wipe out the Flashloader. To do this the patterns at the
64 boundary in flash have been eliminated. The run target logic
that checks for pattern at the bottom of the SRAM is still there,
so an application can cause itself to be debugged by writing
to `run_target` and issue a soft reset. This will cause the
Flashloader to get control.